+++ fabric/src/main/java/net/mehvahdjukaar/moonlight/core/mixins/fabric/FireBlockMixin.java
@@ -1,20 +1,18 @@
 package net.mehvahdjukaar.moonlight.core.mixins.fabric;
 
-import com.llamalad7.mixinextras.sugar.Share;
-import com.llamalad7.mixinextras.sugar.ref.LocalRef;
+import com.llamalad7.mixinextras.injector.wrapoperation.Operation;
+import com.llamalad7.mixinextras.injector.wrapoperation.WrapOperation;
+import com.llamalad7.mixinextras.sugar.Local;
 import net.mehvahdjukaar.moonlight.api.events.IFireConsumeBlockEvent;
 import net.mehvahdjukaar.moonlight.api.events.MoonlightEventsHelper;
 import net.minecraft.core.BlockPos;
 import net.minecraft.core.Direction;
-import net.minecraft.util.RandomSource;
 import net.minecraft.world.level.Level;
 import net.minecraft.world.level.block.BaseFireBlock;
 import net.minecraft.world.level.block.FireBlock;
 import net.minecraft.world.level.block.state.BlockState;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
-import org.spongepowered.asm.mixin.injection.Inject;
-import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 
 
 @Mixin(FireBlock.class)
@@ -25,25 +23,17 @@ public abstract class FireBlockMixin extends BaseFireBlock {
     }
 
 
-    @Inject(method = "checkBurnOut",
-            at = @At(value = "INVOKE",
-                    target = "net/minecraft/world/level/Level.removeBlock (Lnet/minecraft/core/BlockPos;Z)Z",
-                    shift = At.Shift.AFTER))
-    private void afterRemoveBlock(Level level, BlockPos pos, int chance, RandomSource randomSource, int age,
-                                  CallbackInfo ci, @Share("old_state") LocalRef<BlockState> oldState) {
-        var event = IFireConsumeBlockEvent.create(pos, level, oldState.get(), chance, age, Direction.DOWN);
-        MoonlightEventsHelper.postEvent(event, IFireConsumeBlockEvent.class);
-        BlockState newState = event.getFinalState();
-        if (newState != null) level.setBlockAndUpdate(pos, newState);
-    }
-
-    @Inject(method = "checkBurnOut",
+    @WrapOperation(method = "checkBurnOut",
             at = @At(value = "INVOKE",
                     target = "net/minecraft/world/level/Level.removeBlock (Lnet/minecraft/core/BlockPos;Z)Z"))
-    private void beforeRemoveBlock(Level level, BlockPos blockPos, int i, RandomSource random, int j, CallbackInfo ci,
-                                   @Share("old_state") LocalRef<BlockState> oldState) {
-        oldState.set(level.getBlockState(blockPos));
+    private boolean afterRemoveBlock(Level level, BlockPos pos, boolean isMoving, Operation<Boolean> original,
+                                     @Local BlockState oldState, @Local(ordinal = 0, argsOnly = true) int chance,
+                                     @Local(ordinal = 1, argsOnly = true) int age) {
+        var event = IFireConsumeBlockEvent.create(pos, level, oldState, chance, age, Direction.DOWN);
+        MoonlightEventsHelper.postEvent(event, IFireConsumeBlockEvent.class);
+        BlockState newState = event.getFinalState();
+        if (newState != null) return level.setBlockAndUpdate(pos, newState);
+        else return original.call(level, pos, isMoving);
     }
-
 
 }